requirements.txt
PySide6
PySide6-Fluent-Widgets
pyserial
typing